Add CI, half-edge invariant tests, and IO round-trip tests
Three gaps addressed together, because the tests are what make the CI worth
having and the CI is what keeps the tests honest.

**CI.** There was none. GitHub Actions now runs tests, clippy, and a format
check on push and PR. Clippy runs once per feature set — library, `cli`,
`viewer` — since feature-gated code is otherwise never linted, and `-D warnings`
also enforces the crate's existing `#![warn(missing_docs)]`. The viewer has no
tests and needs a GPU, so it only gets `cargo check` to stop it rotting. Every
job checks out `approxum` as a sibling, since morsel depends on it by relative
path; there's a comment on what to do if that repo is private.

**Half-edge invariants** (`tests/halfedge_invariants.rs`). `halfedge.rs` was 721
lines behind 3 tests while every algorithm in the crate traverses it. Ten
universal properties — `twin` is a fixed-point-free involution that reverses its
edge, `next`/`prev` are mutual inverses, half-edges chain head-to-tail, face
cycles close in exactly `face_vertex_count` steps and stay in their face,
boundary flags agree with face pointers, boundary loops close, the vertex
circulator yields each outgoing half-edge once, faces are non-degenerate, every
non-boundary half-edge is in exactly one face cycle, and closed meshes are
consistently wound — asserted across ten fixtures spanning a single triangle,
fans, grids, and two closed solids. Euler's formula checks V − E + F against the
expected χ, which pins the edge count globally in a way no local check does.
All ten passed on the first run, which is the good outcome: the structure was
already sound, and now it stays that way.

**IO round trips** (`tests/io_roundtrip.rs`). PLY, STL and glTF had zero tests
and OBJ had two; nothing tested save → load → same mesh. This found two real
precision bugs:

- OBJ silently narrowed every coordinate to `f32`, because `tobj` defaults to
  32-bit and morsel didn't enable its `use_f64` feature. Now enabled, and the
  now-redundant `as f64` casts are removed so that dropping the feature becomes
  a compile error rather than silent narrowing.
- PLY wrote full `f64` digits under a `property float` header, telling every
  reader to throw half of each coordinate away. Now `property double`; the reader
  already accepted either.

Both formats now round-trip bit-exactly, verified on coordinates chosen not to
survive `f32`. STL stays at `f32` — that is the format's own specification, not
something fixable here — and is tested at that precision, including that its
triangle soup re-welds to the original vertex count. Also covers UV round trips,
that glTF save fails cleanly without leaving a file behind, that unknown
extensions are rejected by name, and that a second round trip costs nothing
further. Every round trip re-checks the half-edge invariants, a reader being a
plausible place to introduce a broken structure.

+++ src/io/obj.rs
@@ -72,19 +72,18 @@ pub fn load_with_uvs<P: AsRef<Path>, I: MeshIndex>(
     for model in &models {
         let mesh = &model.mesh;
 
-        // Extract vertices
+        // Extract vertices. `tobj` is built with `use_f64`, so these are already
+        // `f64` and no cast is wanted — which also means that if that feature is
+        // ever dropped this stops compiling rather than quietly narrowing every
+        // coordinate to `f32`.
         for chunk in mesh.positions.chunks(3) {
-            all_vertices.push(Point3::new(
-                chunk[0] as f64,
-                chunk[1] as f64,
-                chunk[2] as f64,
-            ));
+            all_vertices.push(Point3::new(chunk[0], chunk[1], chunk[2]));
         }
 
         // Extract UV coordinates if present
         if mesh.texcoords.len() == mesh.positions.len() / 3 * 2 {
             for chunk in mesh.texcoords.chunks(2) {
-                all_uvs.push(Point2::new(chunk[0] as f64, chunk[1] as f64));
+                all_uvs.push(Point2::new(chunk[0], chunk[1]));
             }
         } else {
             has_uvs = false;
